Handle ReshapeOp during op lowering
ReshapeOp is seens as memory alias operation.
I tried to use MemRefCastOp to avoid buffer allocation.
Buf MemRefCastOp requires the rank unchanged.

So I still use buffer allocation but place DallocOp after the last use
of the user of ReshapeOp.
The moveBefore() cannot place DeallocOp in the exact position,
So I put it in the end block.

Change-Id: If6e8c93ae3fac05e4e2f4f85e59637ecd47ff88b