Restrict process in python npz_compare
Current npz_compare creates a process for each name at the same time.
For resnet_res_blstm_tensor_all_fp32.npz, 2259 processes are created.

Python 3.8.5 in ubuntu 20.04 prohibits to create so many processes.
And it wastes the system resource to do it.

So restrict the process to cpu cores.

Update to Sam's version.

Change-Id: Ib1b1a3bfcbd825810e3c83d2ad32303e08e0e2f9