Refine fc compressed weight
Current compressed weight pass records each position of tiled position.
The position varies with the actual compressed size so it may generate
a lot of tiled information in mlir file.

Replace it with simpler method - compressed position as the same as
plain data.
TDMA compression does not support stride load, we need to place
the tiled compressed weight in the contiguous memory.

Since the weight (right matrix) is accessed as column-major order, the
tiled compressed weight is stored with this order.

Change-Id: I4d1b5bc8137827c26c7d905eafdec8268cc22b71